Refactor so that SlouchDB is no longer managing a database
SlouchDB is now designed to just delegate database requests (insert,
update, deleted) and database fetches to its delegate and dataSource,
respectively. The client how must provide its own data store object,
which will most likely just be SQLite or some other persistent store.

SlouchDB is now simply a coordinator of local and remote diffs. Any
local mutations on the database should go through SlouchDB, which will
then create a diff entry in the local journal, following that, it will
delegate the mutation request to its delegate, which the client
provides. When a sync is done, "merge" requests are also delegated to
the client. The client now simply provides a persistent database of its
own.

This likely warrants a name change to SlouchJournal or something similar.