Major refactor and reimagining
Instead of requiring clients to encode the data that is being modified
and providing to ChangeTracker. Instead just let them provide opaque
Operation data in a Command type.

ChangeTracker is now really a CommandTracker. Each client can append
commands for a given object id and each client can process the list of
commands as it sees fit. The commands can be anything the client wishes
them to be since the Operation field is just an opaque data type.

This lets clients be super flexible and define whatever their "change"
domain is going to be.