Update JournalManager to delegate its state storing
1. Update JournalManager so that it delegates loading and storing its state to a new JournalManagerStateStoring object.

The previous implementation would store it in a dumb .json file. The problem with that strategy was that if we used
SQLite for the Object History Store, we could use transactions for its changes. We couldn't do that with JournalManager
since multiple clients had the ability to force a save of the .json at any time. With this new method, we can use
SQLite and have it store the state on-demand (and with transactions).

2. With new JournalManagerStateStoring, change the data schema to allow for more than one local identifier.

With this new scheme, we can store as many local identifiers as needed and get the latest one via the localIdentifier()
method.

3. Add resetSyncState() to ChangeTracker.

This will reset the journal states such that it will appear all journals need to be replayed from scratch. This should
not corrupt any data since it will just re-create the database state. This is useful if for some reason the journal
state becomes corrupt.