Fixed subtle bug where journal reader could get stuck
The bug is that if we were unable to read a line of data in a .journal
file, we could sometimes leave the file pointer in the middle of a JSON
fragment. This can happen if the writer is not done writing a full line
yet, but the reader reads part of it. If left in this state, the next
time we try to read a line of JSON, we start in the middle of it instead
of at the start.

The fix is to detect this situation and then adjust the file pointer by
seeking backwards until we reach the first 0x0a newline char. Once we
find that, we re-set the file pointer and start reading lines again from
there. This is only attempted on the first read since subsequent reads
should not encounter this issue.