Fix major bug wherer sync could get stuck
The reason for the sync getting stuck was that startProcessingIfNeeded
was being called even before currentCommandCompletions were processed.
This meant that we ran the risk of missing out on completions if we
started processing the next set of commands too soon. This is likely
what was causing many Fresh Cards sync issues.

This also adds logging delegate methods