When upload error is encountered, simply ignore it instead of blocking the sync process
The old behavior was if we had a problem uploading a file, we bailed on
the sync completely and wouldn't even bother pulling remote changes in.
This would entirely bork the sync until somehow a future attempt at
syncing would miraculously cause the upload on the problem file to work.

This new behavior just plows through and ignores upload failures and
continues with the rest of the steps. We weren't showing the upload
failures to the user anyway, so this is much better behavior since at
least they can sync the remaining files.